+++ dorklab/__init__.py
@@ -13,6 +13,6 @@
 
 """DorkLab - ricerca avanzata e protective dorking con interfaccia grafica."""
 
-__version__ = "1.6.1"
+__version__ = "1.6.2"
 APP_NAME = "DorkLab"
 APP_ID = "dorklab"